Refactor fork controls to own component
