+++ src/components/CollateralTypesTable.tsx
@@ -3,12 +3,13 @@ import { Badge, SortDescriptor, Table, Text, User } from '@nextui-org/react';
 import { addressEq, wadToDec, ZERO } from '@fiatdao/sdk';
 import {
   decodeCollateralTypeId, earnableRateToAPY, encodeCollateralTypeId, encodePositionId,
-  floor2, formatUnixTimestamp, getPositionData, getTimestamp, interestPerSecondToAPY, interestPerSecondToRateUntilMaturity, scaleAndConvertMaturity
+  floor2, formatUnixTimestamp, getPositionData, interestPerSecondToAPY, interestPerSecondToRateUntilMaturity, scaleAndConvertMaturity
 } from '../utils';
 import { chain as chains, useAccount, useNetwork, } from 'wagmi';
 import { useCollateralTypes } from '../state/queries/useCollateralTypes';
 import { useUserData } from '../state/queries/useUserData';
 import useStore from '../state/stores/globalStore';
+import { USE_FORK } from './HeaderBar';
 
 export const CollateralTypesTable = () => {
   const [sortedData, setSortedData] = React.useState<any[]>([]);
@@ -19,6 +20,7 @@ export const CollateralTypesTable = () => {
   const fiat = useStore((state) => state.fiat);
   const setSelectedPositionId = useStore((state) => state.setSelectedPositionId);
   const setSelectedCollateralTypeId = useStore((state) => state.setSelectedCollateralTypeId);
+  const ganacheTime = useStore((state) => state.ganacheTime);
 
   const { data: collateralTypesData } = useCollateralTypes(fiat, chain?.id ?? chains.mainnet.id);
   const { data: userData } = useUserData(fiat, chain?.id ?? chains.mainnet.id, address ?? '');
@@ -86,8 +88,9 @@ export const CollateralTypesTable = () => {
               const borrowRate = interestPerSecondToRateUntilMaturity(interestPerSecond, maturity);
               const borrowRateAnnualized = interestPerSecondToAPY(interestPerSecond);
               const maturityFormatted = scaleAndConvertMaturity(maturity).getTime();
-              const now = Math.floor(getTimestamp() / 1000);
-              const daysUntilMaturity = Math.max(Math.floor((Number(maturity.toString()) - now) / 86400), 0);
+              const now = (USE_FORK ? ganacheTime : new Date()).getTime();
+              const nowSeconds = Math.floor(now / 1000);
+              const daysUntilMaturity = Math.max(Math.floor((Number(maturity.toString()) - nowSeconds) / 86400), 0);
               return (
                 <Table.Row key={encodeCollateralTypeId(vault, tokenId)}>
                   <Table.Cell>
@@ -111,7 +114,7 @@ export const CollateralTypesTable = () => {
                   <Table.Cell>{`${floor2(wadToDec(borrowRateAnnualized.mul(100)))}% (${floor2(wadToDec(borrowRate.mul(100)))}%)`}</Table.Cell>
                   <Table.Cell>{`${floor2(Number(wadToDec(depositedCollateral))).toLocaleString()} ${symbol}`}</Table.Cell>
                   <Table.Cell css={{'& span': {width: '100%'}}}>
-                    <Badge isSquared color={getTimestamp() < maturityFormatted ? 'success' : 'error'} variant='flat' >
+                    <Badge isSquared color={now < maturityFormatted ? 'success' : 'error'} variant='flat' >
                       {formatUnixTimestamp(maturity)}, ({daysUntilMaturity} days)
                     </Badge>
                   </Table.Cell>
