+++ src/components/HeaderBar.tsx
@@ -1,13 +1,13 @@
 import React from 'react';
-import { Badge, Button, Dropdown, Input, Link, Text, Tooltip } from '@nextui-org/react';
+import { Badge, Button, Link, Text, Tooltip } from '@nextui-org/react';
 import { connectButtonCSS, ProxyButton } from './ProxyButton';
 import { ConnectButton } from '@rainbow-me/rainbowkit';
 import { ResourcesModal } from './ResourcesModal';
 import { queryMeta } from '@fiatdao/sdk';
-import { chain as chains, useAccount, useBlockNumber, useNetwork, useProvider } from 'wagmi';
+import { chain as chains, useAccount, useBlockNumber, useNetwork } from 'wagmi';
 import { useFiatBalance } from '../state/queries/useFiatBalance';
 import useStore from '../state/stores/globalStore';
-import { hexValue } from 'ethers/lib/utils';
+import { ForkControls } from './ForkControls';
 
 interface BlockSyncStatus {
   subgraphBlockNumber: number;
@@ -17,82 +17,20 @@ interface BlockSyncStatus {
   message: string;
 }
 
-const SESSION_STORAGE_KEY = 'fiat-ui-snapshotIds';
-
-const SECONDS_PER_DAY = 60 * 60 * 24;
-const SECONDS_PER_WEEK = SECONDS_PER_DAY * 7;
-const SECONDS_PER_MONTH = SECONDS_PER_DAY * 31;
-const SECONDS_PER_YEAR = SECONDS_PER_DAY * 365;
-
-const fastForwardOptions = [{
-  label: 'Forward By 1D',
-  value: SECONDS_PER_DAY
-}, {
-  label: 'Forward By 1W',
-  value: SECONDS_PER_WEEK
-}, {
-  label: 'Forward By 1M',
-  value: SECONDS_PER_MONTH
-}, {
-  label: 'Forward By 1Y',
-  value: SECONDS_PER_YEAR
-}];
-
 export const USE_GANACHE = process.env.NEXT_PUBLIC_GANACHE_LOCAL === 'true';
 export const USE_TENDERLY = process.env.NEXT_PUBLIC_TENDERLY_FORK === 'true';
 export const USE_FORK = ( USE_GANACHE || USE_TENDERLY ) && process.env.NODE_ENV === 'development';
 
-interface SnapshotId {
-  time: number;
-  id: string;
-}
-
 export const HeaderBar = (props: any) => {
   const [showResourcesModal, setShowResourcesModal] = React.useState<boolean>(false);
   const [syncStatus, setSyncStatus] = React.useState<BlockSyncStatus>();
-  const [snapshotIds, setSnapshotIds] = React.useState<SnapshotId[]>([]);
   const {data: providerBlockNumber, refetch} = useBlockNumber();
-  const provider = useProvider() as any;
 
   const fiat = useStore((state) => state.fiat);
-  const getGanacheTime = useStore((state) => state.getGanacheTime);
-  const ganacheTime = useStore((state) => state.ganacheTime);
 
   const { chain } = useNetwork();
   const { address } = useAccount();
   const { data: fiatBalance } = useFiatBalance(fiat, chain?.id ?? chains.mainnet.id, address ?? '');
-
-  const handleFastForward = async (time: number) => {
-    if (!USE_FORK) return;
-    await handleSnapshot();
-    if (USE_GANACHE) {
-      await provider.send('evm_increaseTime', [time]);
-      await provider.send('evm_mine', [{blocks: 1}]);
-    } else {
-      const increaseTimeParams = [hexValue(time)];
-      await provider.send('evm_increaseTime', increaseTimeParams)
-      const increaseBlocksParams = [hexValue(1)];
-      await provider.send('evm_increaseBlocks', increaseBlocksParams);
-    }
-    await getGanacheTime();
-  }
-
-  const handleSnapshot = async () => {
-    const result = await provider.send('evm_snapshot')
-    const newSnapshotIds = [...snapshotIds, {time: useStore.getState().ganacheTime.getTime(), id: result}];
-    setSnapshotIds(newSnapshotIds);
-    window.sessionStorage.setItem(SESSION_STORAGE_KEY, JSON.stringify(newSnapshotIds));
-  }
-
-  const handleRevert = async (snapshotId: string) => {
-    if (!snapshotId) return;
-    await provider.send('evm_revert', [snapshotId])
-    getGanacheTime();
-    const index = snapshotIds.findIndex((item) => snapshotId === item.id);
-    const newSnapshotIds = index > -1 ? snapshotIds.slice(0, index) : [];
-    setSnapshotIds(newSnapshotIds);
-    window.sessionStorage.setItem(SESSION_STORAGE_KEY, JSON.stringify(newSnapshotIds));
-  }
 
   const queryBlockNumber = React.useCallback(async () => {
     if (!fiat) return;
@@ -126,21 +64,13 @@ export const HeaderBar = (props: any) => {
   }, [fiat, providerBlockNumber])
 
   React.useEffect(() => {
-    if (!USE_FORK) return;
-    const resultString = window.sessionStorage.getItem(SESSION_STORAGE_KEY);
-    if (!resultString) return;
-    setSnapshotIds(JSON.parse(resultString));
-  }, [])
-
-  React.useEffect(() => {
     queryBlockNumber();
-    if (USE_FORK) getGanacheTime();
     const timer = setInterval(() => {
       refetch();
       queryBlockNumber();
     }, 10000);
     return () => clearInterval(timer);
-  }, [fiat, queryBlockNumber, refetch, getGanacheTime])
+  }, [fiat, queryBlockNumber, refetch])
 
   return (
     <div style={{ display: 'flex', flexDirection: 'column' }}>
@@ -159,22 +89,7 @@ export const HeaderBar = (props: any) => {
           <Text size='$xs'css={{ alignSelf: 'center'}}>
             {syncStatus?.subgraphBlockNumber}
           </Text>
-          { USE_FORK && (
-            <>
-              <Dropdown closeOnSelect={false}>
-                <Dropdown.Button size='xs' css={{ marginLeft: '3px' }}>{ganacheTime?.toLocaleString().split(',')[0]}</Dropdown.Button>
-                <Dropdown.Menu disabledKeys={['input']} aria-label="Fast Forward" onAction={(e) => handleFastForward(parseInt(e as string))}>
-                  {fastForwardOptions.map((item) => (<Dropdown.Item key={item.value}>{item.label}</Dropdown.Item>))}
-                </Dropdown.Menu>
-              </Dropdown> 
-              <Dropdown closeOnSelect={false}>
-                <Dropdown.Button size='xs' css={{ marginLeft: '3px' }}>Snapshots</Dropdown.Button>
-                <Dropdown.Menu disabledKeys={['input']} aria-label="Snapshots" onAction={(e) => handleRevert(e as string)}>
-                  { snapshotIds.map((item) => (<Dropdown.Item key={item.id}>{`Revert to ${new Date(item.time).toLocaleDateString()}`}</Dropdown.Item>)) }
-                </Dropdown.Menu>
-              </Dropdown>
-            </>
-          )}
+          <ForkControls/>
         </Tooltip>
       </div>
       <div style={{ display: 'flex', justifyContent: 'space-between', padding: 0 }}>
