+++ src/components/PositionsTable.tsx
@@ -11,6 +11,7 @@ import { ethers } from 'ethers';
 import { useCollateralTypes } from '../state/queries/useCollateralTypes';
 import { useUserData } from '../state/queries/useUserData';
 import useStore from '../state/stores/globalStore';
+import { USE_FORK } from './HeaderBar';
 
 export const PositionsTable = () => {
   const [sortedData, setSortedData] = React.useState<any[]>([]);
@@ -21,7 +22,8 @@ export const PositionsTable = () => {
   const fiat = useStore((state) => state.fiat);
   const setSelectedPositionId = useStore((state) => state.setSelectedPositionId);
   const setSelectedCollateralTypeId = useStore((state) => state.setSelectedCollateralTypeId);
-  
+  const ganacheTime = useStore((state) => state.ganacheTime);
+
   const { chain } = useNetwork();
   const { address } = useAccount();
 
@@ -113,8 +115,9 @@ export const PositionsTable = () => {
               const dueAtMaturity = normalDebt.mul(borrowRate).div(WAD);
               const collRatio = computeCollateralizationRatio(collateral, fairPrice, normalDebt, virtualRate);
               const maturityFormatted = scaleAndConvertMaturity(maturity).getTime();
-              const now = Math.floor(getTimestamp() / 1000);
-              const daysUntilMaturity = Math.max(Math.floor((Number(maturity.toString()) - now) / 86400), 0);
+              const now = (USE_FORK ? ganacheTime : new Date()).getTime();
+              const nowSeconds = Math.floor(now / 1000);
+              const daysUntilMaturity = Math.max(Math.floor((Number(maturity.toString()) - nowSeconds) / 86400), 0);
               return (
                 <Table.Row key={encodePositionId(vault, tokenId, owner)}>
                   <Table.Cell>
@@ -154,7 +157,7 @@ export const PositionsTable = () => {
                     }
                   </Table.Cell>
                   <Table.Cell css={{'& span': {width: '100%'}}}>
-                    <Badge isSquared color={getTimestamp() < maturityFormatted ? 'success' : 'error'} variant='flat' >
+                    <Badge isSquared color={now < maturityFormatted ? 'success' : 'error'} variant='flat' >
                       {formatUnixTimestamp(maturity)}, ({daysUntilMaturity} days)
                     </Badge>
                   </Table.Cell>
